requirements.txt
Django==1.10.3
django-allauth==0.29.0
django-nocaptcha-recaptcha==0.0.19
django-social-auth==0.7.28
future==0.16.0
httplib2==0.9.2
jsonfield==1.0.3
MySQL-python==1.2.5
oauth2==1.9.0.post1
oauthlib==2.0.1
pycryptodomex==3.4.3
pyjwkest==1.3.2
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.3.1
requests==2.12.4
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.1.0
